refactor(tree): prepare for new sequence field changeset format
